+++ packages/ui/menu/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pikas-ui/menu
 
+## 3.2.1
+
+### Patch Changes
+
+- fix colors props of item
+
 ## 3.2.0
 
 ### Minor Changes